# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8089
    description: Generated server url
tags:
  - name: Wallet
    description: Service wallet
paths:
  /api/v1/wallet/:
    get:
      tags:
        - Wallet
      summary: Get all wallets
      operationId: findAll
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    put:
      tags:
        - Wallet
      summary: Update wallet
      operationId: updateWallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletUpdateRequest'
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    post:
      tags:
        - Wallet
      summary: Create a new wallet
      operationId: createWallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /api/v1/wallet/id/{id}:
    get:
      tags:
        - Wallet
      summary: Get a wallet by id
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Bad Request
        '404':
          description: Not found
  /api/v1/wallet/{id}:
    delete:
      tags:
        - Wallet
      summary: Delete wallet
      operationId: deleteWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json: {}
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
components:
  schemas:
    WalletUpdateRequest:
      required:
        - idWallet
        - numberDocument
        - phoneNumber
        - typeDocument
      type: object
      properties:
        typeDocument:
          type: string
          description: Type document
          example: 1 = DNI | 2 = RUC | 3 = CE
        numberDocument:
          type: string
          description: Number document
          example: '12345678'
        phoneNumber:
          type: string
          description: Phone number
          example: '123456789'
        email:
          type: string
          description: Email
          example: pedro@gmail.com
        idWallet:
          type: string
          description: Id wallet
          example: 120cf999662f294fc1234567
    Wallet:
      type: object
      properties:
        id_wallet:
          type: string
        typeDocument:
          type: string
        numberDocument:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        registerDate:
          type: string
          format: date-time
        status:
          type: boolean
    WalletRequest:
      required:
        - numberDocument
        - phoneNumber
        - typeDocument
      type: object
      properties:
        typeDocument:
          type: string
          description: Type document
          example: 1 = DNI | 2 = RUC | 3 = CE
        numberDocument:
          type: string
          description: Number document
          example: '12345678'
        phoneNumber:
          type: string
          description: Phone number
          example: '123456789'
        email:
          type: string
          description: Email
          example: pedro@gmail.com